pyproject.toml
[tool.poetry]
name = "chesspal-mcp-engine"
version = "0.2.0"
description = "A Stockfish-powered chess engine exposed as an MCP server using FastMCP. Calculates best moves via MCP tools accessible over SSE (default) or stdio transports using an MCP client library. Part of the ChessPal project."
authors = ["Wilson Urdaneta <wilson.urdaneta@gmail.com>"]
readme = "README.md"
packages = [
    { include = "dylangames_mcp_chess_engine", from = "src" }
]
license = "GPL-3.0-or-later"
repository = "https://github.com/wilson-urdaneta/dylangames-mcp-chess-engine"
keywords = ["chess", "stockfish", "mcp", "chesspal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Games/Entertainment :: Board Games",
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
mcp = {version = "^1.2.0", extras = ["cli"]}
python-chess = "^1.999"
setuptools = "^69.2.0"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic-settings = "^2.8.1"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
black = "^24.2.0"
flake8 = "^7.0.0"
flake8-docstrings = "^1.7.0"
isort = "^5.13.2"
pre-commit = "^3.6.2"
httpx = "^0.28.1"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chesspal-mcp-engine = "dylangames_mcp_chess_engine.main:main"

[tool.black]
line-length = 79
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
max-line-length = 79
extend-ignore = "E203"
docstring-convention = "google"
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:D100,D101,D102,D103"
]
exclude = [".git", "__pycache__", "build", "dist"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "-v"
markers = [
    "integration: marks tests as integration tests (require external dependencies like Stockfish binary)",
]
